refactor: Separating utilites to helper file
